Split parameter.hpp between header and source
