+++ CMakeLists.txt
@@ -48,6 +48,7 @@ add_library(
     source/kettle_internal/io/read_tangelo_file.cpp
     source/kettle_internal/io/statevector.cpp
     source/kettle_internal/io/write_tangelo_file.cpp
+    source/kettle_internal/parameter/parameter.cpp
     source/kettle_internal/simulation/measure.cpp
     source/kettle_internal/simulation/multithread_simulate_utils.cpp
     source/kettle_internal/simulation/operations.cpp
